practice 3 fix+ array

+++ src/module3/practice/Bank/Convertation.java
@@ -7,7 +7,7 @@ public class Convertation {
         this.course = course;
     }
 
-    public double convertToDol(int grvn){
+    public double convertToDol(double grvn){
         return grvn * course;
     }
